refactor(resources): migrate resources package from pkg to internal
